+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Typo fix in marketplace author contact form
 - Added changelog link in the about page
 - Sorting by color key removed from ingredients category settings
+- Making ingredient category modal static
 
 ### Version 7.4
 - Formula making view confirmation dialogs are now static